Kill at boot when storage can't be reached
Right now GitLab tries to check if the disks are reachable, but this is
not it's responsibility.

This gave a couple of options to how Gitaly would do this, and I've
opted to detect at boot if the storage is reachable. A forgotten mount
will be clear this way, and logged without being hidden between lots of
other messages.