Make git hooks self healing
If the git hooks aren't symlinked properly, GitLab authentication and
authorization is skipped completely. So on most Git operations, if these
hooks are missing this commit will try to repair them when triggering an
upload-pack, or receive-pack git action.

The `hooks.SetGitLabHooks()` method is idempotent, meaning it will work
correctly in the case where the hooks are set properly.

Also, setting the hooks like this will allow the Gitaly team to, at some
point, change the location of the Git hooks without migrating them in
one big swoop.

To make sure each repository has the right hooks set up, the hooks are
set when accessing the hooks. Old ones are moved, as this is atomic and
faster. The cleanup RPC has been extended to include the moved hooks.

During tests, hooks are symlinked to the `tmphook/hooks` dir, as else
the test will maintain state in the gitlab-test repository that will
influence the results of other tests.

Part of https://gitlab.com/gitlab-org/gitaly/issues/1226